refactor: use ButtonV2 in auth forms
