+++ packages/shared/src/components/auth/AuthHeader.tsx
@@ -4,7 +4,7 @@ import React, {
   KeyboardEvent,
   ReactElement,
 } from 'react';
-import { Button } from '../buttons/Button';
+import { Button, ButtonVariant } from '../buttons/ButtonV2';
 import ArrowIcon from '../icons/Arrow';
 import { Modal } from '../modals/common/Modal';
 import { ModalHeaderKind } from '../modals/common/types';
@@ -35,7 +35,8 @@ function AuthHeader({
       {onBack && (
         <Button
           icon={<ArrowIcon className="-rotate-90" />}
-          className="mr-2 btn-tertiary"
+          variant={ButtonVariant.Tertiary}
+          className="mr-2"
           onClick={onBack}
         />
       )}
